requirements.txt
ollama==0.3.0
faiss-cpu==1.8.0
numpy==1.24.3
python-dotenv==1.0.1
flask==2.3.2
psycopg2-binary==2.9.7
sqlalchemy==2.0.20